Error Solved

+++ Concepts Of Programming Languages/C/include/Oyun.h
@@ -7,7 +7,7 @@ struct OYUN{
     void (*Basla)(struct OYUN* oyun);
     Kisi* kisiler;
     int* sayilar;
-    float masaParasi; 
+    double masaParasi; 
    
 };
